creat exceptions and unit tests

+++ src/main/java/com/example/criandoapis/services/UserService.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.criandoapis.entities.User;
+import com.example.criandoapis.exceptions.UserNotFoundException;
+import com.example.criandoapis.exceptions.UsernameAlreadyExistsException;
 import com.example.criandoapis.repository.UserRepository;
 
 @Service
@@ -15,11 +17,15 @@ public class UserService {
     private UserRepository userRepository;
 
     public User saveUser(User user) {
+            if (userRepository.existsByUsername(user.getUsername())) {
+            throw new UsernameAlreadyExistsException("Username already exists: " + user.getUsername());
+        }
         return userRepository.save(user);
     }
 
     public User getUserById(UUID id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id)
+        .orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
     }
 
     public List<User> getAllUsers() {
@@ -27,6 +33,19 @@ public class UserService {
     }
 
     public void deleteUser(UUID id) {
+        if(userRepository.existsById(id) == false){
+            throw new UserNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
+    }
+
+    public User editUser(User user){
+        UUID id = user.getId();
+        User existingUser = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
+                existingUser.setUsername(user.getUsername());
+                existingUser.setPassword(user.getPassword());
+        User updatedUser = userRepository.save(existingUser);
+        return updatedUser;            
     }
 }
